Add "has-network" attribute
This allows us to show or hide things based on network status.
Allowable values are "online" or "offline".

An [earlier attempt][1] at this introduced an `is-offline` attribute,
but the disadvantage of this approach is that "online" is an implicit
default state. The approach here puts "online" and "offline" on the same
logical footing, and allows them to be combined with other conditions
more easily.

Note that if a value other than "offline" or "online" is provided to
`has-network` it will always return false. This seemed the most
consistent approach with other matchers (eg. supports, media, etc).

[1]: https://github.com/oddbird/show-when/pull/7